fix activejob initialization sequence
